+++ extensions/reclaim-ai/CHANGELOG.md
@@ -1,5 +1,8 @@
 # reclaim Changelog
 
+## [Fixes] - 2024-09-04
+- Break out separate components and hook for Scheduling Links
+
 ## [Fixes] - 2024-09-03
 - Replace axios with node-fetch